requirements.txt
anthropic==0.49.0
gradio==5.21.0
gradio_client==1.7.2
huggingface-hub==0.29.3
ipykernel==6.29.5
ipython==9.0.2
Jinja2==3.1.6
nest-asyncio==1.6.0
numpy==2.2.3
pandas==2.2.3
python-dotenv==1.0.1
tqdm==4.67.1
